+++ inst/vosondash/server/twitterServer.R
@@ -272,13 +272,25 @@ observeEvent(twitter_view_rvalues$data, {
 
 #### output ---------------------------------------------------------------------------------------------------------- #
 
-output$twitter_collect_token_output <- renderText({
+observeEvent(creds_rv$use_token, {
+  value <- "No token set"
   if (!is.null(creds_rv$use_token)) {
-    createTokenId(creds_rv$use_token)
-  } else {
-    "No token set"
+    value <- createTokenId(creds_rv$use_token)
   }
+  updateTextInput(
+    session,
+    "twitter_collect_token_output",
+    value = value
+  )
 })
+
+# output$twitter_collect_token_output <- renderText({
+#   if (!is.null(creds_rv$use_token)) {
+#     createTokenId(creds_rv$use_token)
+#   } else {
+#     "No token set"
+#   }
+# })
 
 # render twitter collection arguments
 output$twitter_arguments_output <- renderText({
@@ -425,7 +437,7 @@ datatableTwitterData <- reactive({
       col_defs <- gbl_dt_col_defs
       col_defs[[1]]$targets = "_all"
     }
-    DT::datatable(data, extensions = 'Buttons', filter = "top",
+    DT::datatable(data, extensions = 'Buttons', filter = "top", selection = "none",
                   options = list(lengthMenu = gbl_dt_menu_len, pageLength = gbl_dt_page_len, scrollX = TRUE,
                                  columnDefs = col_defs, dom = 'lBfrtip',
                                  buttons = c('copy', 'csv', 'excel', 'print')),