+++ inst/vosondash/ui/youtubeUI.R
@@ -1,5 +1,6 @@
 #### youtube collection tab ------------------------------------------------------------------------------------------- #
 tabItem(tabName = "youtube_collection_tab",
+
         fluidRow(
                 column(
                         width = 3,